Automated_Sentiment/requirements.txt
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
nltk==3.6.3
tensorflow==2.6.0
requests==2.26.0
beautifulsoup4==4.10.0
matplotlib==3.4.3
seaborn==0.11.2
yfinance==0.1.63
alpha_vantage==2.3.1
requests==2.26.0
newsapi-python==0.2.6
joblib==1.0.1
scipy==1.7.1